Splitting fisher allocation into multiple files for readability
