+++ ic/fisher/basic_fisher.py
@@ -1,0 +1,102 @@
+def run_basic_market(initial_values, agent_settings, market_settings, plotting=False, rational=False):
+    u, agent_constraints = agent_settings
+    y, p, r = initial_values
+    w, supply, beta = market_settings
+
+    x_iter = 0
+    prices = []
+    rebates = []
+    overdemand = []
+    agent_allocations = []
+    error = [] * len(agent_constraints)
+    while x_iter <= 100:  # max(abs(np.sum(opt_xi, axis=0) - C)) > epsilon:
+        # Update agents
+        x = update_basic_agents(w, u, p, r, agent_constraints, y, beta, rational=rational)
+        agent_allocations.append(x)
+        overdemand.append(np.sum(x, axis=0) - supply.flatten())
+        for agent_index in range(len(agent_constraints)):
+            constraint_error = agent_constraints[agent_index][0] @ x[agent_index] - agent_constraints[agent_index][1]
+            if x_iter == 0:
+                error.append([constraint_error])
+            else:
+                error[agent_index].append(constraint_error)
+
+        # Update market
+        k, y, p, r = update_basic_market(x, (1, p, r), (supply, beta), agent_constraints)
+        rebates.append([rebate_list for rebate_list in r])
+        prices.append(p)
+        x_iter += 1
+    if plotting:
+        for good_index in range(len(p)):
+            plt.plot(range(1, x_iter+1), [prices[i][good_index] for i in range(len(prices))], label=f"Good {good_index}")
+        plt.xlabel('x_iter')
+        plt.ylabel('Prices')
+        plt.title("Price evolution")
+        plt.legend()
+        plt.show()
+        plt.plot(range(1, x_iter+1), overdemand)
+        plt.xlabel('x_iter')
+        plt.ylabel('Demand - Supply')
+        plt.title("Overdemand evolution")
+        plt.show()
+        for agent_index in range(len(agent_constraints)):
+            plt.plot(range(1, x_iter+1), error[agent_index])
+        plt.title("Constraint error evolution")
+        plt.show()
+        for constraint_index in range(len(rebates[0])):
+            plt.plot(range(1, x_iter+1), [rebates[i][constraint_index] for i in range(len(rebates))])
+        plt.title("Rebate evolution")
+        plt.show()
+        for agent_index in range(len(agent_allocations[0])):
+            plt.plot(range(1, x_iter+1), [agent_allocations[i][agent_index] for i in range(len(agent_allocations))])
+        plt.title("Agent allocation evolution")
+        plt.show()
+    print(f"Error: {[error[i][-1] for i in range(len(error))]}")
+    print(f"Overdemand: {overdemand[-1][:]}")
+    return x, p, r, overdemand
+
+
+def update_basic_market(x, values_k, market_settings, constraints):
+    '''Update market consumption, prices, and rebates'''
+    shape = np.shape(x)
+    num_agents = shape[0]
+    num_goods = shape[1]
+    k, p_k, r_k = values_k
+    supply, beta = market_settings
+    
+    # Update consumption
+    y = cp.Variable((num_agents, num_goods))
+    objective = cp.Maximize(-(beta / 2) * cp.square(cp.norm(x - y, 'fro')) - (beta / 2) * cp.square(cp.norm(cp.sum(y, axis=0) - supply, 2)))
+    # cp_constraints = [y >= 0]
+    # problem = cp.Problem(objective, cp_constraints)
+    problem = cp.Problem(objective)
+    problem.solve(solver=cp.CLARABEL)
+    y_k_plus_1 = y.value
+
+    # Update prices
+    p_k_plus_1 = p_k + beta * (np.sum(y_k_plus_1, axis=0) - supply)
+    for i in range(len(p_k_plus_1)):
+        if p_k_plus_1[i] < 0:
+            p_k_plus_1[i] = 0
+
+    # Update each agent's rebates
+    r_k_plus_1 = []
+    for i in range(num_agents):
+        agent_constraints = constraints[i]
+        if UPDATED_APPROACH:
+            constraint_violations = np.array([agent_constraints[0][j] @ x[i] - agent_constraints[1][j] for j in range(len(agent_constraints[1]))])
+
+        else:
+            constraint_violations = np.array([max(agent_constraints[0][j] @ x[i] - agent_constraints[1][j], 0) for j in range(len(agent_constraints[1]))])
+        r_k_plus_1.append(r_k[i] + beta * constraint_violations)
+    return k + 1, y_k_plus_1, p_k_plus_1, r_k_plus_1
+
+
+def update_basic_agents(w, u, p, r, constraints, y, beta, rational=False):
+    num_agents = len(w)
+    num_goods = len(p)
+    x = np.zeros((num_agents, num_goods))
+    for i in range(num_agents):
+        x[i,:] = update_agent(w[i], u[i,:], p, r[i], constraints[i], y[i,:], beta, rational=rational)
+    # print(x)
+    return x
