+++ ic/main.py
@@ -308,7 +308,7 @@ def run_scenario(data, scenario_path, scenario_name, file_path, method="fisher",
 
     file_name = file_path.split("/")[-1].split(".")[0]
     # data = load_json(file_path)
-    output_folder = f"ic/results/{file_name}_{design_parameters['beta']}_{design_parameters['dropout_good_valuation']}_{design_parameters['default_good_valuation']}_{design_parameters['price_default_good']}_{design_parameters['rebate_frequency']}"
+    output_folder = f"ic/results/{file_name}_{design_parameters['beta']}_{design_parameters['dropout_good_valuation']}_{design_parameters['default_good_valuation']}_{design_parameters['price_default_good']}_{design_parameters['rebate_frequency']}_hardconstraints"
     Path(output_folder).mkdir(parents=True, exist_ok=True)
 
     flights = data["flights"]
